dashboard/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
kafka-python==2.0.2
python-dotenv==1.0.0 